tahieu update v1.0.0 project

+++ FastFood/app/src/main/java/com/tavanhieu/fastfood/adapters/CartAdapter.java
@@ -1,14 +1,20 @@
 package com.tavanhieu.fastfood.adapters;
 
 import android.annotation.SuppressLint;
+import android.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
+
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.squareup.picasso.Picasso;
 import com.tavanhieu.fastfood.R;
+import com.tavanhieu.fastfood.activities.purchase_order.StatusToPayActivity;
 import com.tavanhieu.fastfood.my_class.BuyProduct;
 import com.tavanhieu.fastfood.sqlite_roomdb.MyDatabase;
 
@@ -52,16 +58,26 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.ViewHolderCart
         BuyProduct item = arr.get(position);
         DecimalFormat df = new DecimalFormat("0.00");
         if (item == null) return;
-        holder.imgItem.setImageResource(item.getImage());
+        //Load dữ liệu:
         holder.txtTitle.setText(String.valueOf(item.getTitle()));
         holder.txtNumberOfItem.setText(String.valueOf("x" + item.getNumber()));
         holder.txtPrice.setText(String.valueOf("$" + item.getPrice()));
         holder.txtTotalPrice.setText(String.valueOf("$" + df.format(item.getPrice() * item.getNumber())));
+        //Tải ảnh:
+        Picasso.get().load(item.getImage()).centerInside().resize(200, 200).into(holder.imgItem);
 
         holder.imgCancel.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                MyDatabase.getMyInstances(holder.itemView.getContext()).myDao().delete(item.getId());
+                new AlertDialog.Builder(holder.imgItem.getContext())
+                        .setTitle("Notification")
+                        .setMessage("Do you want to remove this product?")
+                        .setNegativeButton("No", (dialog, which) -> {
+                        })
+                        .setPositiveButton("Yes", (dialog, which) -> {
+                            MyDatabase.getMyInstances(holder.itemView.getContext()).myDao().delete(item.getId());
+                            Toast.makeText(holder.itemView.getContext(), "Remove " + item.getTitle(), Toast.LENGTH_SHORT).show();
+                        }).show();
             }
         });
     }
